pass to double precision

+++ test/test_pagerank.cpp
@@ -4,6 +4,8 @@
 #include "vector.hpp"
 #include <cstdio>
 
+// #define PRECISION double
+
 void clean_pagerank_files(const graphee::Properties& props){
     for(int i = 0; i < props.nslices; i++){
         for(int j = 0; j < props.nslices; j++){
@@ -45,11 +47,11 @@ BOOST_AUTO_TEST_CASE( test_dense_web )
      */
     pagerank.compute_pagerank(100);
     long n=0;
-    float score_sum=0;
-    graphee::Vector<float> vec(&props);
+    double score_sum=0;
+    graphee::Vector<double> vec(&props);
     for(uint64_t slice_i=0; slice_i<props.nslices; slice_i++){
         vec.load("test_pagerank_web_pr_dvecslc_"+std::to_string(slice_i)+".gpe");
-        for(float score : vec){
+        for(double score : vec){
             score_sum+=score;
             n++;
         }
@@ -89,12 +91,12 @@ BOOST_AUTO_TEST_CASE( test_smallGraph )
      */
     pagerank.compute_pagerank(10);
     long n=0;
-    float score_sum=0;
-    graphee::Vector<float> vec(&props);
-    float expected[] = {0.21495,0.15189,0.03953,0.26713,0.22387,0.10260};
+    double score_sum=0;
+    graphee::Vector<double> vec(&props);
+    double expected[] = {0.21495,0.15189,0.03953,0.26713,0.22387,0.10260};
     for(uint64_t slice_i=0; slice_i<props.nslices; slice_i++){
         vec.load("test_smallGraph_pr_dvecslc_"+std::to_string(slice_i)+".gpe");
-        for(float score : vec){
+        for(double score : vec){
             std::cout<<n<<"\t"<<score<<std::endl;
             score_sum+=score;
             BOOST_CHECK(abs(score-expected[n])<0.00001);
@@ -136,12 +138,12 @@ BOOST_AUTO_TEST_CASE( test_oneEdge )
      */
     pagerank.compute_pagerank(3);
     long n=0;
-    float score_sum=0;
-    graphee::Vector<float> vec(&props);
-    // float expected[] = {0.21495,0.15189,0.03953,0.26713,0.22387,0.10260};
+    double score_sum=0;
+    graphee::Vector<double> vec(&props);
+    // double expected[] = {0.21495,0.15189,0.03953,0.26713,0.22387,0.10260};
     for(uint64_t slice_i=0; slice_i<props.nslices; slice_i++){
         vec.load("test_oneEdge_pr_dvecslc_"+std::to_string(slice_i)+".gpe");
-        for(float score : vec){
+        for(double score : vec){
             std::cout<<n<<"\t"<<score<<std::endl;
             score_sum+=score;
             n++;
@@ -172,11 +174,11 @@ BOOST_AUTO_TEST_CASE( test_smallGraph_sum_columns )
   filenames.push_back("test/ressources/test_sum_columns.txt.gz");
   adjacency_matrix.load_edgelist(filenames);
 
-    graphee::DiskVector<graphee::Vector<float>> out_bounds = std::move(graphee::DiskVector<graphee::Vector<float>>(&props, "ob", 0.));
+    graphee::DiskVector<graphee::Vector<double>> out_bounds = std::move(graphee::DiskVector<graphee::Vector<double>>(&props, "ob", 0.));
 out_bounds.dmat_columns_sum(adjacency_matrix);
-float n = 0;
+double n = 0;
 for(int slice = 0 ; slice < props.nslices ; slice++){
-      graphee::Vector<float>& vec = out_bounds.get_slice(slice);
+      graphee::Vector<double>& vec = out_bounds.get_slice(slice);
       for(int i = 0;i<vec.get_lines();i++){
         BOOST_CHECK_EQUAL(vec.at(i),n);
         n+=1;
